pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "0.21.0"
pyTelegramBotAPI = "^4.12.0"
psycopg2-binary = "^2.9.6"
peewee = "^3.16.2"
fastapi = "^0.97.0"
aiohttp = "^3.8.4"
redis = "^4.5.5"
qrcode = "^7.4.2"
uvicorn = "^0.22.0"
crc16 = "^0.1.1"
flagsmith = "3"
google-cloud-storage = "^2.13.0"
firebase-admin = "^6.2.0"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.app:start"